Add process payment method

+++ src/test/java/org/example/PaypalStrategyTest.java
@@ -9,7 +9,7 @@ public class PaypalStrategyTest {
     @Test
     public void testCheckout() {
         PaypalStrategy paypalStrategy = new PaypalStrategy();
-        double amount = 10.0;
+        int amount = 10;
         String result = paypalStrategy.checkout(amount);
         assertEquals("Checkout with Paypal: $" + amount, result, "Should return Checkout with Paypal: $10.0");
     }
